requirements.txt
#python -m pip install --index-url=https://bcms.bloomberg.com/pip/simple blpapi
pandas==2.0.3
plotly==5.14.0
sqlalchemy==2.0.6
mysqlclient==2.1.1
et_lib==2.6.1
kaleido==0.1.0.post1
workalendar==17.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
python-pptx==0.6.21
openpyxl==3.1.2
holidays==0.38
yfinance==0.2.32

